Create the recompress library.
This will allow other code to use the same jpeg-recompress logic.
